Bug fixes
There seems to have been a bug with not cleaning up the array for each bubble removal on replay. I think this is a fix. Also, though a bubble has completed it's animation cycle it still seems to need to be stopped for it to be seen as not animating.